Mockup-Klassen entfernt. Dafür MockBleBroadcaster aufgebohrt.
