--- examples/Darp.Ble.Examples.Unix/Mockup/BMFactory.cs
@@ -1,13 +1,0 @@
-using Microsoft.Extensions.Logging;
-
-namespace Darp.Ble.Examples.Unix.Mockup;
-
-public sealed class BMFactory : IBleFactory
-{
-    public string Name { get; set; } = "Mock";
-
-    public IEnumerable<IBleDevice> EnumerateDevices(ILogger? logger)
-    {
-        yield return new BMDevice(Name, logger);
-    }
-}
